requirements.txt
fastapi
uvicorn
google-adk
google-genai
python-dotenv

google-cloud-pubsub
google-auth
